requirements.txt
spacy
streamlit
pdfminer.six
